msi interrupt

+++ kernel/interrupt.cpp
@@ -1,3 +1,9 @@
+/**
+ * @file interrupt.cpp
+ *
+ * 割り込み用のプログラムを集めたファイル．
+ */
+
 #include "interrupt.hpp"
 
 // #@@range_begin(idt_array)
@@ -5,12 +11,10 @@ std::array<InterruptDescriptor, 256> idt;
 // #@@range_end(idt_array)
 
 // #@@range_begin(set_idt_entry)
-void SetIDTEntry(
-  InterruptDescriptor& desc,
-  InterruptDescriptorAttribute attr,
-  uint64_t offset,
-  uint64_t segment_selector
-) {
+void SetIDTEntry(InterruptDescriptor& desc,
+                 InterruptDescriptorAttribute attr,
+                 uint64_t offset,
+                 uint16_t segment_selector) {
   desc.attr = attr;
   desc.offset_low = offset & 0xffffu;
   desc.offset_middle = (offset >> 16) & 0xffffu;
@@ -21,7 +25,7 @@ void SetIDTEntry(
 
 // #@@range_begin(notify_eoi)
 void NotifyEndOfInterrupt() {
-  volatile auto end_of_interrupt = reinterpret_cast<int32_t*>(0xfee000b0);
+  volatile auto end_of_interrupt = reinterpret_cast<uint32_t*>(0xfee000b0);
   *end_of_interrupt = 0;
 }
 // #@@range_end(notify_eoi)
